x/alphabet/types/message_buy_letter_test.go
package types

import (
	"testing"

	"alphabet/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgBuyLetter_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgBuyLetter
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgBuyLetter{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgBuyLetter{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
